move serde_as trait wrapper alongside backing fn
